UT3 Scorpion Missing Collision and Explosion Sound Fix
This gives the UT3 Scorpion it's impact sounds (which at this time can only really be heard from weapon damage that throws the vehicle around as it doesn't seem to take impact damage from walls or vehicles), and gives it the correct explosion sound on death.  UAX file needs to be uploaded too but how?