UT3 Manta Closer to UT3 Impact Damage
In UT3 the Manta (and actually most vehicles) takes little to no damage on impact, usually around 1 for a wall and 4 for an enemy vehicle or so,  while this value isn't quite perfect, it's close to what UT3 is with the Manta only taking damage while near full speed and only 1-3 usually for a wall.  It might be the damage is reduced too much with this but anything higher results in damage from barely any speed.