UT3 Hellfire SPMA Headlights Adjustment
We've got a slight problem with the headlights, they are meant to come from the wheel plates which turn and change position while moving forward or backward causing floating headlights....I'm guessing there isn't a way to fix this so the question is which do we want, headlights that are lined up when the vehicle is sitting still (probably best considering SPMA is usually deployed when you see it) or do we want them to line up while moving forward?  These values are for sitting still, I haven't come up with ones for moving yet but will if I need to.