Manta: Engine Flames WIP, Help Needed
DO NOT MERGE, I need help solving an issue first

Looking at a mod Manta I got the trail effects to work but the Manta has reduced jump height now and I can't figure out why and there is nothing I can see that would cause it.   I also am aware the UT3 Manta doesn't have streamers but considering the Manta is a fast vehicle I'd like to keep that for myself at least.   It also looks like I may have found what's needed to fix the Manta's high pitch sound issue but not sure how to work with it.

It's almost as if the Manta is now heavier, even getting in it doesn't jump up as much, it does seem this is the only negative impact so far which if necessary could probably tweak default properties to get something that works.